修改了两个接口

+++ DataEntity/DataEntities.cs
@@ -27,7 +27,7 @@ namespace MirRemakeBackend.DataEntity {
         public readonly short m_level;
         public readonly int m_upgradeExperienceInNeed;
         public readonly IReadOnlyList<ValueTuple<ActorUnitMainAttributeType, int>> m_mainAttributeList;
-        public readonly int m_mainAttributePointNum;
+        public readonly short m_mainAttributePointNum;
         public DE_CharacterData (DO_Character charDo) {
             m_upgradeExperienceInNeed = charDo.m_upgradeExperienceInNeed;
             m_mainAttributeList = new List<ValueTuple<ActorUnitMainAttributeType, int>> (charDo.m_mainAttributeArr);
